Desktop/Projets et realisations/ALEX/requirements.txt
# ALEX - Assistant IA avec RAG - Dependencies
flask>=2.3.3
taipy>=3.1.0
streamlit>=1.28.0
ollama>=0.1.7
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0
sentence-transformers>=2.2.2
pypdf>=3.17.0
docx2txt>=0.8
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0